backend/server/requirements.txt
fastapi[all]==0.109.1
alembic==1.13.2
pydantic==2.7.1
fastapi-sqlalchemy==0.2.1
psycopg2-binary==2.9.9
celery==5.3.6
celery[redis]==5.3.6
python-dotenv==1.0.0
loguru==0.7.2
requests==2.32.3
docker==7.1.0
psutil==5.9.8
pysqlite3-binary==0.5.2
pyyaml==6.0.1
